feat(selection): add opts.includeDeprecated
This is mainly so we can keep using the latest npm-pick-manifest,
which previously included deprecation-skipping as a breaking change.

With this patch, we can flip a switch for the semver-major change on
the npm side of things whenever, and the api for pacote will remain
the same unless the flag is passed in explicitly.